+++ native/common/include/jp_numbertype.h
@@ -32,6 +32,7 @@ public:
 	virtual~ JPNumberType();
 
 	JPMatch::Type findJavaConversion(JPMatch& match) override;
+	virtual void getConversionInfo(JPConversionInfo &info) override;
 } ;
 
 #endif // _JPNUMBERTYPE_H_